Changes are all about gson module isa.
1. uses `kotlin-reflect` dependency now isa.

2. object instance stay the same after deserialization isa.

3. var of sealed class inside data class can be serialized/deserialized isa.

4. even if have for ex. 3 sealed class nested and you wanna conversion
just annotate the root superclass of them and conversion will be done for you isa.

5. coding improvements so that instead of builing gson on every to/from json
calls it's only done once isa.

6. OPS still some minor issues for the above special use case we don't
check for superclass fields like gson does so we need to implement that,
also need cleaner approach by tring all constructors not the first one
for deserialization process, ALSO there is room for more improvements
in the notes so check it out isa.